major changes, mostly repeating busy time and saving settings
I made it so busy time can repeat if the user wants and fixed errors that came with that. I made the week calendar display all hours in the user's day instead of fixed times. I made user settings be saved into and loaded from a file. I fixed the errors with empty boxes on the settings menu. I improved some display sorting algorithms and fixed some errors that occur when a task is removed due to lack of available time. I tried and failed to make some buttons on the keyboard have the same effect as some buttons on the window.